requirements.txt
streamlit
opencv-python
pytesseract
pillow
numpy
tesseract
streamlit-cropper